+++ src/main/java/com/bitbank/entities/UsersEntity.java
@@ -37,6 +37,9 @@ public class UsersEntity {
 
 	@Column(name = "password", columnDefinition = "VARCHAR(256)")
 	private String password;
+	
+	@Column(name = "secret_mfa", columnDefinition = "VARCHAR(256)")
+	private String secretMfa;
 
 	@ManyToMany(fetch = FetchType.LAZY)
 	@JoinTable(name = "users_roles", joinColumns = @JoinColumn(name = "users_id"), inverseJoinColumns = @JoinColumn(name = "roles_id"))