+++ src/main/java/com/bitbank/controllers/AuthController.java
@@ -4,9 +4,6 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.validation.Valid;
-
 import org.modelmapper.ModelMapper;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -40,6 +37,8 @@ import com.bitbank.utils.JwtUtils;
 import com.bitbank.utils.RandomStringGenerator;
 
 import io.swagger.v3.oas.annotations.tags.Tag;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.validation.Valid;
 
 @RestController("AuthControllerV1")
 @RequestMapping("/api/v1/auth")
@@ -85,7 +84,8 @@ public class AuthController {
 	 */
 	@PostMapping("/register")
 	@ResponseStatus(HttpStatus.CREATED)
-	public ResponseEntity<MessageResponse> post(@Valid @RequestBody UsersPostDTO usersDTO) throws RoleOrAuthorityNotFoundException {
+	public ResponseEntity<MessageResponse> post(@Valid @RequestBody UsersPostDTO usersDTO)
+			throws RoleOrAuthorityNotFoundException {
 
 		String username = usersDTO.getUsername();
 		if (Boolean.TRUE.equals(getEnableSubscription())) {
@@ -99,7 +99,7 @@ public class AuthController {
 			RolesEntity customerRole;
 			customerRole = rolesService.show(ERole.ROLE_CUSTOMER);
 			rolesEntity.add(customerRole);
-			
+
 			// Prepare for secret MFA
 			RandomStringGenerator randomStringGenerator = new RandomStringGenerator();
 			String generatedRandomString = randomStringGenerator.getRandomString();
@@ -142,8 +142,7 @@ public class AuthController {
 		String jwt = jwtUtils.generateJwtToken(authentication);
 		// Get the expiry at value
 		String expiryAt = jwtUtils.getExpiryDateFromJwtToken(jwt).toString();
-
-		// Get the autorithies
+		// Get the autorities
 		List<String> roles = jwtUtils.getAuthorities(authentication);
 
 		return ResponseEntity.ok(new JwtResponse(jwt, expiryAt, roles));