+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 + Fix deprecated CORS
 + Add column `secret_mfa` to database
 + Add utility to generate a random string
++ Store secret when save a new user
 
 ## 1.0.0
 + Refactor Semaphore