+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 + Add utility to generate a random string
 + Store secret when save a new user
 + Add V2 controller for login
++ Add `findById` method for user
 
 ## 1.0.0
 + Refactor Semaphore