+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 + Store secret when save a new user
 + Add V2 controller for login
 + Add `findById` method for user
++ Refactor `findById` method for user
 
 ## 1.0.0
 + Refactor Semaphore