+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Next version
+# 1.1.0
 + Upgrade dependencies
 + Fix deprecated CORS
 + Add column `secret_mfa` to database